+++ src/lib.rs
@@ -53,19 +53,18 @@
 //! ```
 //!
 //! The resulting crate must provide an opt-in "rt" feature and depend on these crates:
-//! `bare-metal` v0.2.x, `cortex-m` v0.5.x, `cortex-m-rt` >=v0.6.5 and `vcell` v0.1.x. Furthermore
+//! `cortex-m` v0.7, `cortex-m-rt` >=v0.6.13 and `vcell` v0.1.x. Furthermore
 //! the "device" feature of `cortex-m-rt` must be enabled when the "rt" feature is enabled. The
 //! `Cargo.toml` of the device crate will look like this:
 //!
 //! ``` toml
 //! [dependencies]
-//! bare-metal = "0.2.0"
-//! cortex-m = "0.5.8"
+//! cortex-m = "0.7"
 //! vcell = "0.1.0"
 //!
 //! [dependencies.cortex-m-rt]
 //! optional = true
-//! version = "0.6.5"
+//! version = "0.6.13"
 //!
 //! [features]
 //! rt = ["cortex-m-rt/device"]